+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -321,7 +321,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT8WF7C7FQ6A3HEYJR6F5RZZ"
+        "01KT8Z5A1GSG1VM8MVZXGQRKRN"
     )
     r"""The unique identifier of the data source"""