+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KT8WF7CA7FQTWNGNYJ4XZ5D5"
+        "01KT8Z5A1K61QVM9BDWEKZ6RS4"
     )
     r"""The unique identifier of the data source"""